refactor: consolidate normalize_vllm_version and remove redundant MTEB scripts
Move normalize_vllm_version to config_manager to eliminate duplication across
three dashboard pages. Remove run-mteb-missing-tests.sh and its documentation
as all tests are now consolidated in the main test script.
