+++ automation/test-execution/dashboard-examples/vllm_dashboard/pages/1_📊_Client_Metrics.py
@@ -21,7 +21,7 @@ from plotly.subplots import make_subplots
 
 # Add parent directory to path for config_manager import
 sys.path.insert(0, str(Path(__file__).parent.parent))
-from config_manager import DashboardConfig
+from config_manager import DashboardConfig, normalize_vllm_version
 
 # Set global Plotly template
 if "plotly_white_light" not in pio.templates:
@@ -45,30 +45,6 @@ logging.basicConfig(
     handlers=[logging.StreamHandler(sys.stdout)],
 )
 logger = logging.getLogger(__name__)
-
-
-def normalize_vllm_version(version_string):
-    """Normalize vLLM version for display.
-
-    Maps the current RHAIIS vLLM version to a user-friendly display name.
-    Future versions are displayed as-is.
-
-    Args:
-        version_string: Raw vLLM version from metadata
-
-    Returns:
-        "RHAIIS_3.4" for current version, original string otherwise
-    """
-    if not version_string or version_string == 'unknown':
-        return 'unknown'
-
-    # Current RHAIIS 3.4 version
-    if version_string == '0.18.0+rhaiv.7':
-        return 'RHAIIS_3.4'
-
-    # All other versions (past or future) display as-is
-    return version_string
-
 
 # Page config is set in Home.py for multipage apps
